Async Filter API (#494)
The Filter API is refactored to be fully asynchronous.  Filter API methods such as #onXxxxRequest and #onXxxxResponse are now required to return a `CompletionStage<FilterResult>`. The `FilterResult` encapsulates the message to be forwarded and carries orders (such as close the connection).  The semantics are that the forward
does not take place until the completion stage completes.   The context provides factory methods for creating `FilterResult` objects.

Signed-off-by: kwall <kwall@apache.org>