Move zero-acks Produce request detection earlier
Why:
We want to know earlier when the request is a zero-ack Produce request
because this is a special case where there is no response expected. We
want to prevent Filters from sending responses towards the client in this
case and, in future, avoid storing the correlationId in memory for the
purposes of re-ordering messages.