Flush inbound channel on deferred read complete
Why:
An integration test was failing non-deterministically when the
MetadataResponse handling went async. The client never got a
response.

This was also masked in the KrpcFilterIT because the tests that
tried to exercise the deferring code were not really going async.
Thee RequestResponseMarkingFilter did do some work in another thread
but then blocked while closing it's executor, which meant that the
work was completed before the future was returned to the FilterHandler.
So the deferring path wasn't hit.