Chain the read/write futures when we are deferring
Why:
whenComplete just does the work in it's action, since we go on to
create a future in the action, nothing will be delayed if that future
is itself deferred. So the whole queue of whenCompletes will complete
when the first deferred future completes. We need to compose them
together instead.

I also made it flush when a deferred future completed in case there
was further async work enqueued, it might be better to forward it on
rather than wait for other work to complete before flushing.