Add integration tests for short-circuit response re-ordering
To enable testing of this feature we extend the test KafkaClient to stamp
each Response with a sequenceNumber to track what order they were received
in. Then we can assert that the client received a response for message A
before it received a response for message B.

We also need to extend the MockServer to be able to drop requests to simulate
the zero-acks produce request case where there is no corresponding response
from the broker.