Introduce FilterContributorContext to FilterContributor
Why:
We want to enable Filter Authors to access some resources and metadata
at Filter construction time. The first resource is the channel's executor
so that a Filter can schedule work on the event loop. Maybe they have
fast-executing background tasks that don't require their own pool.

Future examples of things that could come in via the FilterContributorContext are:
1. virtual cluster name
2. other virtual-cluster scoped components like for example a shared HttpClient
3. further executors, maybe kroxy should offer a configurable pool for blocking
work that could be shared by Filters.