Add ApiVersionsService to KrpcFilterContext
Why:
Filters can choose to hijack the connection to send their own extra
requests to the upstream broker. To do this safely we need to know what
versions of the RPC the upstream supports. This service uses the
ApiVersionsFilter to record the upstream versions, but is capable of
hijacking the connection if required. This enables us to use the new
API even if we are intercepting the very first ApiVersions request
(a redundant ApiVersions request would be made). We can also tolerate
cases where the client hasn't sent an ApiVersions request.

The ApiVersionsServiceImpl becomes responsible for the intersection logic,
so that it can be re-used in both cases:
1. intercepting the ApiVersionsResponseData during client-server version
negotiation.
2. sending it's own ApiVersions request if one has not been intercepted
yet.