Tolerate multiple messages being read from client during connection
Enqueue further messages handled by KafkaProxyFrontendHandler while it is in
the CONNECTING or CONNECTED state before it has received the OUTBOUND_ACTIVE
signal from the outbound channel. Then write them all to the outbound when it
becomes active and null out the queue.

Why:
It is possible for clients to send multiple messages to the broker immediately.
In this case, even though we try to disable autoread of the inbound channel we
can still read multiple messages from the socket and have to handle multiple
messages while we are establishing the connection to upstream.