Refactor Contributor
Why:
The contributor interface has this awkward contract of returning null from
it's component methods if the shortName is not supported. We can use an
Optional to make it clear at a glance from the API that a contributor does
not have to supply something. Also we contain this logic in one place, after
we obtain an InstanceFactory for a shortName we don't have to keep passing
the shortName in again. Also the InstanceFactory class will be easier to
extend to indicate if the configuration is required or not.