Refactor FilterContributor
Why:
The Contributor set of interfaces and classes have gotten complex and
we think we could have a simpler interface to explain to Filter Authors
by reducing the Cardinality of contributors and removing the Contributor
interface. This will be at the cost of having some boilerplate code and
duplication.

Benefits are:
1. since each FilterContributor only deals with one type they
don't have to deal with mapping the name to the relevant configClass
and the relevant instance building method.
2. the BaseContributor is removed from the equation so we don't have
to tell users how to integrate with that.
3. we can specialise the FilterContributor to take a construction
context without necessarily having to do that in the other implementations