SafeInvoker returns failed stage if delegate returns illegal null
Because we are calling into potentially third-party code we cannot guarantee
that the Filter will return a non-null stage despite our highest hopes. So the
SafeInvoker now centralises that check so that client code can rely on invokers
obtained using `FilterInvokers::from` not returning null.

Note: We may wish to return to this and move the work into the generated code
to further optimise the code. We would potentially be able to generate a fixed
error completion stage and save creating a new object on failure.