Log at all levels to null-appender in tests
Why:
We are sometimes using a pattern where we check if a log level is enabled for
a logger before we construct the logs. This avoids some un-necessary method
calls when we aren't logging at that level. However this means we are pinged
by sonar for not covering those lines, and rightly, we want to at least know
that our logging code isn't going to throw null pointer exceptions etc.