Remove getTypeName, use the Contributor classname instead
Instead of the contributor declaring a String and then the framework matching
that against the type field from configuration, we will instead support the
type matching:

1. the full class name
2. for non-inner, non-local, non-anonymous classes the simple class name

Why:
This way we will not have to explain custom aliases.

We only support top-level classes being identified by simple class name to
support the pattern where inner Contributors are all named 'Contributor'. We
could alternatively specify that the Contributor class name must be unique and
support using the simple name instead for all Contributors, regardless of whether
they are a member class.