Use service class to identify Contributor to load from
Why:
The user isn't really interested in the Contributor, they just want
to get at a particular Filter. So identifying the Contributor using
it's own classname would be a quirk users would have to understand.
It may be more natural for them to declare they want Filter XYZ using
the class name or simple class name of XYZ.