Parameterize the Filter type produced by FilterFactory
Why:
By introducing the parameter we can enforce that the type
returned by filterType and the type of the object returned from
createFilter must be the same. Which makes the authors life easier.
Also at a glance FilterFactory<Void> and FilterFactory<ConfigType>
looked odd, like they are the types produced by the factory.