Deprecate generic Context/Contributor/ContributorManager
Why:
We found that managing the generics was challenging and presented
more complexity to Filter Authors, so the FilterFactory no longer
extends the generic Contributor. These classes are now used only
to load micrometer hooks and the cluster network address providers.
We can simplify things further by removing this classes.