Remove kek from Kms#decrypt
Why:
We want to encourage KMS implementors to include all data they require to
deserialize as part of the edek, including the immutable reference
to the KEK used to encrypt it. Some implementations like AWS already include
this reference within the ciphertext output when you ask the service
to wrap some data.