Rotate to a new DEK when the old one is exhausted
Work done with the key context is now synchronized such that only one
thread will encrypt OR update the future in the cache when rotation is
required. If the key was destroyed while a thread is waiting for the
lock it can retry the encrypt as the stage in the cache should have been
atomically updated by a previous encrypt call.