Add attempt limit to encryption
If a batch of records cannot be encrypted because a DEK has been
exhausted it will retry, chaining onto the new DEK completionstage.
Attempts limit currently hardcoded to 3.

Harder to prove is that it works with multiple threads. A retry will
be prompted if the DEK is rotated by another thread while this thread
is waiting for the lock on the KeyContext.