Release prefix buffer after use
Why:
We should release the buffer in preparation for a real buffer pool
implemenation. Decided it was better to copy the prefix into a byte[]
on heap since this thing will have a long lifetime of it's own, rather
than keep a ref to the allocated buffer and release it when the context
is destroyed.