Short-circuit when records to encrypt empty, tidy up
Why:
We had some code only allocating buffers if the max size would be non-zero.
With the introduction of the wrapper the encrypted parcel size will be at
least 2 bytes even with an empty body so we should always allocate a buffer.
Short-circuiting also means we don't have worry about the empty case when
calculating the max buffer sizes.