Implement null-value handling
Clients can push a null record value to Kafka to tombstone a key in a compacted topic.
We don't want to break this behaviour so we should not modify a null record value in
the Encryption Filter.

However this is in conflict with the current design of the encryption schema. If the
user has push a null record-value and a non-empty set of Headers, and the proxy decides
it must encrypt the headers, then we would write that into the record-value forwarded to
Kafka. This would break the tombstoning logic.

For now we prohibit this edge case, exploding in flames if the value is null but there
are headers we want to encrypt.

In a subsequent PR we could decide to use a header tactically if the record-value is
null and stuff the wrapped parcel in there.