Release instruction should use rebase/merge strategy.
When merging the release PR, it is important to use the *rebase/merge* strategy.

If this is not done, the SHA of the release tag won't align with the SHA of the commit history of main.  You'll see a message like this in the Git Hub UI when viewing the the release tag. 

> This commit does not belong to any branch on this repository, and may belong to a fork outside of the repository.

Signed-off-by: Keith Wall <kwall@apache.org>