Ensure we can recover if an EDEK decryption fails
Previously if the cached future failed it would have remained cached
forever in the failed state and meant nothing could be decrypted for
that EDEK. What we want to happen is, if an EDEK encoded into a wrapped
parcel cannot be decrypted (due to KMS outage etc) then the fetch request
should fail and subsequent requests should be able to succeed. The EDEK
decryption should be attempted again on those future requests.

Using caffeine aligns it with the EDEK cache and gives us the useful
property that there should be only one pending future in flight per key.

Note: the testing required is ugly, Sam commented on this for the EDEK
PR too. It's tough because the InBandKeyManager knows the parcel format
and we don't have test utilities to independently generate examples of
parcels yet. If we had the ability to generate serialized record examples
we could avoid going through an encrypt cycle with a real KMS implementation
just to get properly serialized records.