encryption filter: generalise exception type
We foresee a class of exceptions where encryption/decryption operations
fail due to some logical reason. There is not a failure as such, like
a connection error to the KMS or your everyday NullPointerException that
has crept into the encryption code, but it should still point at an exceptional
case.

One example would be a request to encrypt more records than the maximum
allowed per Data Encryption Key. That request could fail fast because we cannot
satisfy the request.