Simplify test
Why:
There was a race condition where the test code fired a second request
while a mapping was being asynchronously removed from the cache due
to a failed completion stage. For now, let's just test that the failure
behaves as we expect and revisit the abstractions later to see if theres
some way to separate these behaviours and make them easier to test.