Add vault metrics counting operation attempts and outcomes
Why:
Each KMS will likely need slightly different metrics depending on
which operations are supported and how they are implemented. Maybe
it takes multiple API calls to implement DEK generation for a particular
KMS etc.

Adds metrics:
* 'kroxylicious_kms_vault_request_attempt_total' with tag 'operation' to count total
  attempted requests to Vault.
* 'kroxylicious_kms_vault_request_outcome_total' with tags 'operation', 'outcome', 'failure'
  to count the outcome for every attempt.

Example tag combinations emitted for the outcomes:
* successful EDEK decryptions: operation=decrypt_edek,outcome=success,failure=
* failed alias resolve due to unknown alias: operation=resolve_alias,outcome=failure,failure=not_found
* failed DEK creation due to exception: operation=create_dek,outcome=failure,failure=exception

Note that this granularity is targeted so that we can create the Meters immediately and begin
emitting zeros as soon as the application starts so that applications can more easily alert on
change rates. So things like status codes aren't included as we don't want to emit a metric for
every possible HTTP status code. We could lean on logging for drilling into the failures or
add other counters for these more granular metrics.

Signed-off-by: Robert Young <robeyoun@redhat.com>