Do not retry on unknown entity exception
Why:
Currently we rely on the existence of the key to be the signal
to encrypt. If we retry unknown key exception in the KMS we will
add a big delay for the case where the topic is to be un-encrypted.
This could be fixed by making the filter aware of which topics
should be encrypted, which is still under discuession.

Signed-off-by: Robert Young <robeyoun@redhat.com>