Change the scope of envelope encryption resources
1. KMS is now application-wide, all EnvelopeEncryptionFilters for this
Filter configuration will use the same KMS instance (need to decide who
is responsible for KMS thread safety and document it)
2. KeyManager and KekSelector are now per-event-loop. Multiple filters
associated with the same eventloop will share these resources.

Still TODO is to add some concept of DEK allocation so that KeyManager
asks for a DEK that can be used for N encryptions. This should allow
an application-wide DEK allocator to share the same DEK across multiple
event loops as the stateful encryptor and fine-grained encryption counting
will happen per-event-loop.

Signed-off-by: Robert Young <robeyoun@redhat.com>