Implement DekAllocator
DekAllocator is asked to provide a DekPair for N encryption operations. It
generates Deks on-demand and will hand out references to the same Dek while
there are remaining operations. It is assumed all of those allocated operations
are consumed by the caller.

This is intended to be an application-wide instance shared by all instances
of the EnvelopeEncryptionFilter. The client should allocate large blocks
of encryption operations and cache encryption resources such that they do
not need to call this class often.

As it is intended to be shared we have to ensure it's threadsafe, we lean
on caffeine again to ensure we have a bulkhead to only make one request to
the KMS per key at a time. Then for rotation logic we synchronise on a
context class to ensure the key is only invalidated one time.

Signed-off-by: Robert Young <robeyoun@redhat.com>