Switch back to the channel executor after deferred stage
Why:
If a Filter instance completes it's future from an uncontrolled
thread (they can complete the future how they like), we want to
switch back to the channel loop for the next Filter. We offer
the FilterFactory the eventloop executor at construction time with
javadoc saying it's safe to mutate the Filter state from that
executor (enabling the Filter Author to schedule background or timed
tasks that mutate filter state). This does not hold true if the Filter
work can be driven by uncontrolled threads.

Signed-off-by: Robert Young <robeyoun@redhat.com>