Refactor InBandKeyManager#encrypt to pass MemoryRecords
Why:
We are preparing to make the encrypt aware of the batches within a
MemoryRecords instance and preserve the batches along with their
metadata. We need to change encrypt to take MemoryRecords and return
a CompletionStage<MemoryRecords> so that the batch-preservation can
be handled while we encrypt all the records.

Signed-off-by: Robert Young <robeyoun@redhat.com>