Refactor InBandKeyManager#decrypt to operate on MemoryRecords
Why:
We are preparing to make the decrypt aware of the batches within a
MemoryRecords instance and preserve the batches along with their
metadata. We need to change decrypt to take MemoryRecords and return
a CompletionStage<MemoryRecords> so that the batch-preservation can
be handled while we decrypt all the records.

Signed-off-by: Robert Young <robeyoun@redhat.com>