Refactor BatchAwareMemoryRecordsBuilder to track if it has been closed
Why:
With the new direct passthrough ability where we can hand it an existing
batch and have it written to the buffer, we can no longer exclusively
rely on the state of the builder variable to track whether we've finished
working with the BatchAwareMemoryRecordsBuilder. It can happen that the
last builder in the variable is closed by the time we call build(). Instead
we will add a closed variable. Since the builder is marked NonThreadSafe
we do not consider async calls to build/close.

Signed-off-by: Robert Young <robeyoun@redhat.com>