Vault uses FilterFactoryContext to obtain TLS configuration
Why:
This will enable users to supply alternative sources of trust/keys
for Vault using configuration that matches how we configure upstream
and downstream TLS for the proxy. We route the configuration to the
FilterFactoryContext and ask for a client SSLContext.

Note:
I've intentionally hardcoded in some keys/certs for vault integration
testing as using our existing keytool based utility we can't easily
convert to PEM form without some code like we do in the integration tests.
In future we might want to look at using bouncycastle for our cert/key
manipulation.

Signed-off-by: Robert Young <robeyoun@redhat.com>