Redundantly cancel timeout future
Why:
If the filter completed it's future exceptionally, then the thenApplyAsync
block that cancelled the timeout future was never called. By using a whenComplete
we will cancel the timeout on the success and failure path. It's also safe to
redundantly cancel it if the failure was being triggered by the timeout future
itself.

Signed-off-by: Robert Young <robeyoun@redhat.com>