Extract usage counter
Why:
To improve the readability of Dek by shuffling some responsibilities off it.

Dek remains responsible for pulling together the reference counting and
encrypt counting and encapsulating the secret key material so only it's
Encryptors/Decryptors have access to it. The reference counting logic is
hidden away to be understood independently.

Signed-off-by: Robert Young <robeyoun@redhat.com>