Enable Filters to check if they are currently in dispatch thread
Filters working with asynchronous code may need to be sure that they are
executing in the filter thread where we know it is safe to mutate
filter member state. For example if we make an async HTTP request and have
a CompleteableFuture that has it's completion driven by the HTTP io thread,
we may want to update a Filter's state safely on the dispatch thread.

The `eventLoop()` method of FilterFactoryContext is deprecated and replaced with
`filterThreadExecutor()` because we felt the naming is too netty-centric
and we have tried to hide this detail from the Filter implementations as much
as we can. We do make guarantees that a Filter instance is associated with,
and driven by, a single thread. We call this the Filter Thread.

The FilterThreadExecutor exposes `boolean isInFilterThread()` to ask directly
and also provides a method `completedOnFilterThread` to take a CompletionStage
and switch to the FilterThread to drive the completion if required.

Signed-off-by: Robert Young <robeyoun@redhat.com>