envelope encryption: filter selects topics to encrypt
Why:
Previously we relied on the KMS response when resolving aliases to
determine if encryption should be applied. We take the existance of
the key to mean that we should encrypt. This has some negative
implications. Like if most of the topics do NOT have a key then we
will hit the KMS on every produce request going through the filter
for those unencrypted topics. Also, there are more places where
misconfiguration could mean writing unencrypted data to a topic we
really hoped would be encrypted (mismatched key name in vault and
kek selector).

By making the filter aware of which topics we require encryption
for, we can make it an exceptional case that the key is not found
in the KMS. This logic may change when we come to using features
other than topic name for key selection, but it will likely always
be useful to be able to configure which topics need some amount of
encryption applied so we can short-circuit for unencrypted topics.
We also reduce the cases where unencrypted data could be unexpectedly
written because a configuration mismatch and key missing in the KMS
would result in exceptions rather than unencrypted data forwarded to
the cluster.

Topic selection is permissive by default. So all topics would be
selected for encryption.

If any names or prefixes are configured for inclusion, then the topic
must  match an inclusion rule to be selected.

If any names or prefixes are configured for exclusion, then the topic
must be included and not match any exclusion rules to be selected.

Exclusion beats inclusion, so (include["mytopic"], exclude["mytopic"])
would result in mytopic not being selected for inclusion

Signed-off-by: Robert Young <robeyoun@redhat.com>