Create TopicNameRouter and use it with existing mechanism
Why:
We want to protect the KMS from being hammered asking for aliases
for topics we know we never want to encrypt. To do this we want the
user to be able to configure how to encrypt topics with an exact name
match or with a prefix match.

The Router is built to route from topicName to a destination (kek selector
at first but will likely change as we add things like configurable cipher).
The router doesn't have knowledge of what it's routing to it just has to
route each input topic to exactly one destination. It currently depends
on the destinations classes equals/hashcode when routing so that is can
build the map `Map<D, Set<String>> route(Set<String> topicName)` using
destination as key. This is a little awkward but could be used to
support batched operations if the KMS had support for batching alias resolution.

The rules of routing are:
1. exact-topic-name matches are the highest priority rule
2. the same exact-topic-name cannot be routed to multiple destinations
3. prefix-matches are lower priority that exact-topic-name
4. longer prefixes are higher priority than shorter prefixes
5. the same prefix cannot be routed to multiple destinations
6. inclusion beats inclusion, so it's okay for one configuration to say
  "include prefix a, but exclude prefix ab", then another configuration to
  say "include prefix ab"
7. there is a default destination which is selected if nothing else matches

This means we can use the TopicNameRouter to unambiguously route each
topic name to a destination. Exceptions will be thrown if we attempt to
create ambiguous routes.

Signed-off-by: Robert Young <robeyoun@redhat.com>