Add test for parcel binary compatibility
We need tests to prove that Kroxylicious can read older binary versions
of Parcels. We also need tests to prove that it can write older binary
versions. Comparing the output bytes is a blunt tool and will be fiddly
to track down the reason for failures, but we must always be able to decode
those bytes that we forwarded to Kafka to be stored forever, so starting
and ending with the bytes is worth it to prove it really can.

The tests check:
1. an input Record, serialized with specified encryption configuration and
parcel version produces expected bytes
2. the serialized bytes can be deserialized to produce the expected record
features (Parcel deserializes in terms of value and headers only).

Note that input features may differ from deserialized features becase we can
opt to not serialize headers or values despite them being present in the record.

The tests will immediately fail when a new ParcelVersion is added to the enum,
requiring a new exemplar to be added for the new version to all existing
YAML exemplars. In this case the test will display the serialized base64 in
the failure message (if it can), which could be added to the exemplar file
when you are happy with the output.

Signed-off-by: Robert Young <robeyoun@redhat.com>