Add hardcoded maximum frame size
Why:
Currently when a client is configured to use TLS and it connects to
a plain proxy listener we interpret it as a Kafka frame and read the
first 4 bytes as an int that contains the length of the frame.
Rejecting frame sizes larger than 100 Mebibytes happens to catch.
The first byte is the TLS handshake ContentType, 22. This means
the first four bytes are going to compresi an int greater than 22 << 24
(369098752 in decimal). So our max frame size should blow up on any TLS
handshake.

Signed-off-by: Robert Young <robeyoun@redhat.com>