Add integration test checking encrypt v1 AES cipher deserializable
Why:
We must always be able to parse/decrypt old data (assuming we still
have the KEK referenced by the EDEK). This test is to prove that we
can decrypt binary examples injected into the upstream Kafka. We
check an encrypted non-null value, encrypted null value and that
additional headers on the record reach the client.

Signed-off-by: Robert Young <robeyoun@redhat.com>