Fix flaky unit test
Why:
The test wants to check that our test cipher produces different
output for different topic names. The cipher rotated all the bytes
by a fixed amount determined by the topic name hashcode. Meaning
there was a ~0.0039 chance of the same rotation byte being picked
and causing the test to fail. This change varies the rotation byte
by using the characters of the topicname to modify the rotation for
each index.

Signed-off-by: Robert Young <robeyoun@redhat.com>