Cache unknown alias resolutions for 30 seconds
Why:
On the encryption path, for every Produce Request we look for a KEK for every
topic in the request. If we have obtained a KEK already we would hit our caches
and everything is fine. If there is nothing cached, which is the case when there
is no matching KEK in the KMS, then we ask the KMS to resolve an alias. This means
we hammer the KMS in the case where we do not want to encrypt the data going through.

This change adds a fixed 30 seconds of caching for a not-found-alias. Meaning we
do not call out to the KMS again to resolve alias X and should therefore take the
pressure off. Note this could be amplified by a large number of topics in a production
system so it might be good to have a pessimistic default.

Note this is a stopgap because we think the Filter should not be using the KMS response
to dictate whether we encrypt, so we want to make the Filter aware of which topics
_require_ encryption.

Signed-off-by: Robert Young <robeyoun@redhat.com>