Change AES GCM cipher to require a 256bit key
The original design and recommendations were for a 256bit key but our
CipherSpec declared it was 128bit. Then it's algorithm, AES_256/GCM/NoPadding,
didn't enforce the length so it happily used a 256bit key from the KMS.

This change aligns the CipherSpec naming on 256 and updates the algorithm
to enforce the key size.

Signed-off-by: Robert Young <robeyoun@redhat.com>