Bump EncryptionVersion to 2 and warn users about dead alpha format
Also increased WrapperVersion to V2 and leaves an unsupported V1 which
throws if it's ever used. This allows us to mark that the None AADSpec is
supported since V1.

Why:
We decided that since the serialized data in 0.4.0 didn't align with the
design doc that it was unsupported and we would break backwards compatibility,
guaranteeing backwards compatibility from 0.5.0. If any users stumble on this
they could raise a ticket for us to reimplement v1.

This doesn't cover implementing a robust upgrade process as the encryption
version is not user configurable. We will have to add this when we next modify
the protocol.

Signed-off-by: Robert Young <robeyoun@redhat.com>