pages builds triggered by dependabot merge failed
Why:
The build on push to main has a GITHUB_TOKEN with readonly
permissions and a secret source of Dependabot. It cannot push
the updated ghpages branch to the repo. This is a workaround
to instead run a trigger build on main push and then trigger
a separate workflow which will have the Actions GITHUB_TOKEN
permissions.

Signed-off-by: Robert Young <robeyoun@redhat.com>