Extract kafka record batch transformation into module
Creates a kroxylicious-kafka-message-tools module for classes that are
purely interested in manipulating the classes in kafka-clients independent
of any proxy behaviour.

This should allow user filters to use the behaviour but enable them to
use a kroxylicious-kafka-message-tools version independent of the kroxylicious
version.

The first behaviour added to this module is record transformation. That is
we want to enable filters to mutate the records contained in fetch and produce
requests while preserving how the records are batched inside their MemoryRecords
along with all the batch metadata.

Signed-off-by: Robert Young <robeyoun@redhat.com>