Add Indexed Int Set and implementations
Preparing to support range aware port per broker implementation.
IndexedIntSets represent sets of integers where each value in the
set has a unique index in the range [0, size). In practise this
will be done by ordering the values ascending and then using the
position in the resulting list as the index.

Signed-off-by: Robert Young <robeyoun@redhat.com>