Add RangeAwarePortPerNodeClusterNetworkAddressConfigProvider
This new implementation introduces the concept of Node Id Ranges
and enables you to configure multiple such ranges for your virtual
cluster. Enabling a deterministic mapping of node ids to proxy
ports even when there are large gaps between the ranges.

For example if I configure a virtual cluster with:

```
nodeStartPort: 9000
nodeIdRanges:
  - name: brokers
    range:
      startInclusive: 4
      endExclusive: 6
  - name: controllers
    range:
      startInclusive: 1
      endExclusive: 3
```
Then the ports will be mapped as follows:
- node 1 -> port 9000
- node 2 -> port 9001
- node 4 -> port 9002
- node 5 -> port 9003

Validations:
1. endExclusive must be greater than startInclusive
2. range names must be unique within nodeIdRanges
3. ranges must be distinct, a node id must be within exactly 1 range

Why:
Currently there are numerous drawbacks to using the PortPerBroker
exposition scheme. If the lowest broker id is N and the highest is M
then we must expose M-N deterministically map each node id to a port.
So if you have two nodes with ids 100000 and 200000 you need 100000
ports mapped so that we can say startPort+0 is node 100000 and
startPort+100000 is node id 200000.

Also with the advent of KRaft we now may have to contend with
two distinct and evolving sets of node ids, a controller pool and
a broker pool. If there is a significant gap between them we have
the same problem as above.

This enables us to represent clusters with highly variable node ids
using a more compact set of ports.

Note it still does not solve other issues, like some users regulurly
introduce new nodes and remove old ones in their target cluster as part
of upgrades.  Leading to eternal growth of node ids. To handle this in
a better way we think the Proxy would need to be able to discover the target
topology.

Co-authored-by: Keith Wall <kwall@redhat.com>
Co-authored-by: Sam Barker <sam@quadrocket.co.uk>

Signed-off-by: Robert Young <robeyoun@redhat.com>