Add FilterDispatchExecutor
* Deprecates ScheduledExecutorService FilterFactoryContext#eventLoop()
* Adds FilterDispatchExecutor FilterFactoryContext#filterDispatchExecutor()

Why:
The framework guarantees that their Filter methods will be dispatched by a
single thread (the event loop for the Filter instance's channel), which
gives Filter Authors this nice property that they can mutate members of
the Filter safely, as Filter invocation is single-threaded.

We also give the Filter Auther an asynchronous API using Java's CompletionStage,
so it's possible to move work to arbitrary threads. If the Author needs to
change Threads then they no longer have this convenient thread safety
guarantee and would have to co-ordinate access to the members.

To fix this we added FilterFactoryContext#eventLoop() so that the
Author can use the executor to switch back to the Filter Dispatch Thread
and safely access members. The fact it exposes a ScheduledExecutorService
is inflexible if we want to add to it and the naming is implementation
specific. We would prefer to avoid the netty naming.

We also at some stage broke our guarantee that Filter methods would be
dispatched on the Filter Dispatch Thread and have implemented a test to
try and verify what thread we are on. This test obtained the name of the
thread at dispatch time and passed it to the client for assertion. This
is roundabout and a lot of code for checking what we want. It would be
simpler if we had some way to assert the invariant that we must be in
the Filter Dispatch Thread.

To address the naming and flexibility of eventLoop we move to the more
domain-specific naming of filterDispatchExecutor() and add our own
Interface that extends ScheduledExecutorService.

To enable Authors and our tests to detect if execution is happening on
the Filter Dispatch Thread we have a boolean isInFilterDispatchThread()
method. To make it convenient to switch back from an arbitrary thread
we offer <T> CompletionStage<T> switchToFilterDispatchThread(CompletionStage<T> stage);

Signed-off-by: Robert Young <robeyoun@redhat.com>